Create a status file and add function to change status of checkbox
